Fixed benchmarking bugs in primitive HashMap test.
Both IntIntHashMapTest and LongIntHashMapTest failed to reset test start time.

The time values for intmapGetTime,mapGetTime,intmapRemoveTime,and mapRemoveTime
all used the start time value from mapPutTime. This caused the following to be
always/guarenteed to be true:
mapPutTime < intmapGetTime < mapGetTime < intmapRemoveTime < mapRemoveTime

thus not actually testing the following asserts:
  assertTrue("'get' too slow", intmapGetTime <= mapGetTime);
  assertTrue("'remove' too slow", intmapRemoveTime <= mapRemoveTime);

In addition the results of the test were being printed before the actual test
ran.

Changed code to reset time to System.nanoTime() prior to each for loop, and
moved time calculation and result output to below the test for loop.

I also improved the output of the test info, by including the warmup status
on benchmark start.