Fixed benchmark timing inconsitancies.
While running unit tests, the benchmark timing tests of the primitive maps
would periodically fail on a semi-regular basis after an 'ant clean all'
for the first run of a test.

Added a heap allocation stage prior to running tests and added calls to the
GarbageCollector prior to each timing to reduce the chances that the test
thread could be starved while the Garbage Collector was running mid-test.

Appears to fix all timing skews that I was detecting on my local-machine.