Fix TempJarCache's Multi-User Bug (Reported by Martin Hegedus)
It turns out that Java's File mkdir() only makes the directory writable for the current user,
I have missed this fact. Great catch.

1. Fix TempJarCache.isInitialized(): Return false if not successfully initialized.

It merely returned if it has passed 'initSingleton()' and ignored the staticInitError.

2. Fix TempFileCache pattern of determining the temp base directory

We cannot just use a static directory name, due to the multi user environment
and user write permissions on File.mkdir().

IOUtil has a new 'getTempDir(..)' methods, which iterates through integers [000000-999999]
until a writeable directory could be found or created.

TempFileCache initializes this temp base dir in the static block ensuring
the value is final for the JVM / ClassLoader.

Updated comments/docs in TempFileCache.