Bug 845: Add support for one big-fat jar file [java classes plus all native 'os.and.arch' libraries]
JNILibLoaderBase.addNativeJarLibsImpl(..):

If the modules's jar file contains the folder 'natives/<os.and.arch>/'
we assume a big-fat jar and attempt to load all native libraries from the same.

The test for above folder is performed via the class ClassLoader's getResource(..)
and is considered inexpensive.

If the folder exists and native libraries could be loaded, the method returns successfull.

Otherwise, the 'slim' jar file is attempted to be loaded, even if such folder exist.