Allow Ignore in config to ignore unnamed structs.
Augments the behavior of the Ignore keyword to
ignore when gluegen skips the emission of unnamed
structs. These structs are usually unnamed because
we've created opaque types for them, so we know
they're not going to be emitted and want to be
able to suppress the warning on a case-by-case
basis.