Bug 1363: Java 11: Refine 'java.home.dir' and 'java.lib.dir.platform' setup for traditional layout and JEP 220(Java 9+)
JEP 220 states it is now optional to use the <arch> subfolder in 'lib' to store native libraries,
i.e. 'lib/<arch>/libjava.so' can be flattened to 'lib/libjava.so'.
Further the jre subfolder is no more used according to the JEP 220, however, it can be used.

Therefor we scan for 'java.home.dir' in the following order:
- if '<java.home>/../jre' exists, i.e. we are within the 'jre' folder, use '<java.home>/..'.
- otherwise assume <java.home> reflect the flattened actual base folder and use it as is.

We scan for 'java.lib.dir.platform' in the following order:
- if exists <java.home>/jre/lib/i386/libjava.so -> <java.home>/jre/lib/i386
- if exists <java.home>/lib/i386/libjava.so     -> <java.home>/lib/i386
- defaults to flattened <java.home>/lib

This way we keep the historical arch information for each platform
and stay most flexible for any SDK build layout.