Bump cpptasks.jar, commit 19c854265956942497dc89444a9ee84f18383671 (Support native windows dll-names)
Add support for Windows dynamic library target, i.e. NAME.dll instead of libNAME.dll.

This is now required when linking a GLUENAME1.dll to TOOLNAME1.dll,
where GLUENAME1.dll is loaded via System.loadLibrary() and TOOLNAME1.dll required to be found.
Might be a change in the current ming64 toolchain.

Before this change, the library name 'libTOOLNAME1.so' was written into `libGLUENAME1.so`
and hence could not be found (manually renamed to a dll name).

Now we can also drop the manual rename of libNAME1.so -> NAME1.dll, finally :)