Roll back MacOS Min Version Requirement to 10.5 from 11.0 (macosx-version-min)
MacOSX compiler configuration.

2023-05-04 Actively supported architectures are 'x86_64' and 'arm64'.

XCode 14.2 on a MacOS 12.6.5 x86_64 (mac-mini) machine
is capable to produce fat-lipo code
for MacOS >= 10.5 for 'x86_64' and 'arm64'.
The MacOS SDK 11.3 has been re-added from a previous XCode installation.
However, alternative source is <https://github.com/phracker/MacOSX-SDKs>.

Only for using C++ and hence the C++ STL,
the minimum version is 10.7 due to clang++ libc++.
This is currently the case for JOAL's bundled openal-soft,
as well as for JOGL's optional oculusvr-sdk (disabled).